feat(feed-you): refactor views and add feed you page
